agentapp/pyproject.toml
[project]
name = "agentapp"
version = "0.1.0"
description = "Minimal remote DevOps agent app"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "pyyaml>=6.0.2",
  "uvicorn>=0.35.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
